Move tests to 2 separate files
